web_app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
ultralytics>=8.2.0
Pillow>=10.0.0
